Fix various compiler warnings on MSVC
I tried to avoid casts whereever it was possible, but for some cases it wasn't.

+++ BlockPos.h
@@ -142,8 +142,8 @@ public:
 	NodeCoordHashed(const BlockPos &pos) : NodeCoord(pos) { rehash(); }
 	NodeCoordHashed(const NodeCoord &coord) : NodeCoord(coord) { rehash(); }
 	void rehash(void) { m_hash = NodeCoord::hash(); }
-	unsigned hash(void) { rehash(); return m_hash; }
-	unsigned hash(void) const { return m_hash; }
+	size_t hash(void) { rehash(); return m_hash; }
+	size_t hash(void) const { return m_hash; }
 	bool operator==(const NodeCoordHashed &coord) const { if (m_hash != coord.m_hash) return false; return NodeCoord::operator==(coord); }
 	void operator=(const NodeCoord &coord) { NodeCoord::operator=(coord); rehash(); }
 	bool operator<(const NodeCoordHashed &coord) { return m_hash < coord.m_hash; }
